update observations to float32

+++ api/dataset_test.go
@@ -573,7 +573,7 @@ func getValidCantabularResponse(name, label string) cantabular.StaticDatasetQuer
 						},
 					},
 				},
-				Values: []int{100},
+				Values: []float32{100.0},
 			},
 		},
 	}
